refactor(lists): extract lists filter options into a hook
